+++ internal/pkg/util/fs/files/passwd.go
@@ -83,12 +83,8 @@ func Passwd(path string, home string, uid int, customLookup UserGroupLookup) (co
 		}
 		if entry.Uid() == uid {
 			userExists = true
-			// If user already exists in container, update their homedir
-			gid32, err := safecast.Convert[uint32](entry.Gid())
-			if err != nil {
-				return nil, err
-			}
-			lines[i] = makePasswdLine(entry.Username(), uid32, gid32, entry.Info(), homeDir, entry.Shell())
+			// If user already exists in container, rebuild their passwd info preserving their original shell value
+			lines[i] = makePasswdLine(pwInfo.Name, pwInfo.UID, pwInfo.GID, pwInfo.Gecos, homeDir, entry.Shell())
 			break
 		}
 	}